refactor(account): rename `identity` module to `account` and update related configurations
- Removed `WebConfig` class and moved CORS configuration to `SecurityConfig`.
- Implemented dedicated `PasswordConfig` for encoder bean setup.
- Updated package names from `modules.identity` to `modules.account` across the codebase.
- Refined JWT and password hashing components for consistency.
