--- src/main/java/com/gresk/infrastructure/config/WebConfig.java
@@ -1,17 +1,0 @@
-package com.gresk.infrastructure.config;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-public class WebConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**") // Permite todos los endpoints
-                .allowedOrigins("http://localhost:3000") // Permite tu Frontend de v0
-                .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
-                .allowedHeaders("*")
-                .allowCredentials(true);
-    }
-}
