+++ src/main/java/com/gresk/modules/account/infrastructure/security/BcryptPasswordHasher.java
@@ -1,6 +1,6 @@
-package com.gresk.modules.identity.infrastructure.security;
+package com.gresk.modules.account.infrastructure.security;
 
-import com.gresk.modules.identity.application.port.out.PasswordHasherPort;
+import com.gresk.modules.account.application.port.out.PasswordHasherPort;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
